権限周りを整理

+++ app/Http/Controllers/ProductsController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Resources\ProductResource;
 use App\Product;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Str;
 use Symfony\Component\HttpFoundation\Response;
@@ -14,6 +15,8 @@ class ProductsController extends Controller
 
     public function index()
     {
+        Gate::authorize('view', 'products');
+
         $products = Product::paginate();
 
         return ProductResource::collection($products);
@@ -27,6 +30,7 @@ class ProductsController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('edit', 'products');
 
         $product = $request->only('title', 'description', 'image', 'price');
 
@@ -40,6 +44,8 @@ class ProductsController extends Controller
      */
     public function show(Product $product)
     {
+        Gate::authorize('view', 'products');
+
         return new ProductResource($product);
     }
 
@@ -52,6 +58,8 @@ class ProductsController extends Controller
      */
     public function update(Request $request, Product $product)
     {
+        Gate::authorize('edit', 'products');
+
         $product->update($request->only('title', 'description', 'image', 'price'));
 
         return response(new ProductResource($product), Response::HTTP_ACCEPTED);
@@ -65,6 +73,9 @@ class ProductsController extends Controller
      */
     public function destroy(Product $product)
     {
+        Gate::authorize('edit', 'products');
         Product::destroy($product);
+
+        return \response(null, Response::HTTP_NO_CONTENT);
     }
 }
